refactor(hr): hoist domain-agnostic primitives to common::shared_models (#473)
Move 7 domain-agnostic HR primitive types (I18nText, FlexibleText, IdNameObject, CodeNameObject, PaginatedResponse<T>, CatalogItem, LocalizedLabel) from hire/hire/common_models.rs to a new crate-root module common::shared_models. hire imports back from there; the 6 sibling domains can now opt-in to typed i18n instead of side-dependending into the hire subtree or falling back to serde_json::Value. Precedent: okr/okr/v2/common/models cross-leaf dedup (#336).

Non-breaking: serialization shape is byte-identical, and hire::hire::common_models keeps a #[deprecated] named re-export of the 7 types for one cycle so existing full-path imports keep resolving; removal targeted at the next breaking window (0.19). rustc does not currently surface deprecation warnings for pub use re-exports at consumer sites, so the alias removal will be grep-driven — documented inline. All internal call sites migrated (incl. tests/hire_tests.rs). Verified: clippy --all-targets (all-features + no-default-features) clean, 1952 tests pass, cargo doc clean.
