+++ crates/openlark-hr/src/hire/hire/v1/interview_record/list.rs
@@ -12,7 +12,7 @@ use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use std::collections::HashMap;
 
-use crate::hire::hire::common_models::{I18nText, IdNameObject};
+use crate::common::shared_models::{I18nText, IdNameObject};
 
 /// `ListRequestBody`。
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
