+++ crates/openlark-hr/src/hire/hire/v1/interview_registration_schema/list.rs
@@ -12,7 +12,7 @@ use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use std::collections::HashMap;
 
-use crate::hire::hire::common_models::CatalogItem;
+use crate::common::shared_models::CatalogItem;
 
 /// `ListRequestBody`。
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
