+++ crates/openlark-hr/src/hire/hire/v1/talent_object/query.rs
@@ -12,7 +12,7 @@ use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use std::collections::HashMap;
 
-use crate::hire::hire::common_models::I18nText;
+use crate::common::shared_models::I18nText;
 
 /// 获取人才字段请求。
 ///
