+++ crates/openlark-hr/src/hire/hire/v1/talent_tag/list.rs
@@ -13,7 +13,7 @@ use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use std::collections::HashMap;
 
-use crate::hire::hire::common_models::I18nText;
+use crate::common::shared_models::I18nText;
 
 /// `ListRequest` 请求。
 #[derive(Debug, Clone)]
